+++ src/platform/monitoring/DowntimeNotification/reducer/index.js
